requirements.txt
pyyaml>=6.0
qrcode>=7.4
requests>=2.28
